requirements.txt
argcomplete==1.10.3
backports.zoneinfo==0.2.1
beautifulsoup4==4.8.2
blis==0.7.9
catalogue==2.0.8
certifi==2022.9.24
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
compressed-rtf==1.0.6
confection==0.0.3
cycler==0.11.0
cymem==2.0.7
docx2txt==0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1-py3-none-any.whl

et-xmlfile==1.1.0
extract-msg==0.28.7
Flask==2.2.2
fonttools==4.38.0
fuzzywuzzy==0.18.0
gunicorn==20.1.0
idna==3.4
IMAPClient==2.1.0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
langcodes==3.3.0
language-tool-python==2.7.1
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.3
murmurhash==1.0.9
nltk==3.7
numpy==1.21.6
olefile==0.46
openpyxl==3.0.10
packaging==21.3
pandas==1.3.5
pathy==0.7.1
pdfminer==20191125
pdfminer.six==20191110
Pillow==9.3.0
preshed==3.0.8
pycryptodome==3.15.0
pydantic==1.10.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-pptx==0.6.21
python-rake==1.5.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
rake-nltk==1.0.6
regex==2022.10.31
requests==2.28.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.12.0
sklearn==0.0.post1
smart-open==5.2.1
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.4.3
spacy-legacy==3.0.10
spacy-loggers==1.0.3
SpeechRecognition==3.8.1
srsly==2.4.5
textract==1.6.5
thinc==8.1.5
threadpoolctl==3.1.0
tqdm==4.64.1
two-lists-similarity==0.0.5
typer==0.7.0
typing-extensions==4.1.1
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
wasabi==0.10.1
Werkzeug==2.2.2
xlrd==1.2.0
XlsxWriter==3.0.3
zipp==3.10.0
